django/abcd/requirement.txt
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
cycler==0.10.0
DateTime==4.3
Django==3.0.6
Flask==1.1.2
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib==3.2.1
mysql-connector==2.2.9
mysqlclient==1.4.6
numpy==1.18.4
Pillow==7.1.2
pymongo==3.10.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.9
Werkzeug==1.0.1
zope.interface==5.1.0
